+++ package.json
@@ -48,6 +48,7 @@
     "react-country-region-selector": "^3.4.0",
     "react-datetime": "2.16.3",
     "react-dom": "16.12.0",
+    "react-dropzone": "^7.0.0",
     "react-google-maps": "9.4.5",
     "react-redux": "^7.2.8",
     "react-router-dom": "5.1.2",
