+++ src/views/discount/DiscountCodes.js
@@ -16,6 +16,7 @@ import {
   withStyles,
   Button,
 } from "@material-ui/core";
+import MatEdit from '../MatEdit';
 
 class DiscountCodes extends React.Component {
   constructor(props) {
@@ -113,6 +114,24 @@ class DiscountCodes extends React.Component {
         type: "used",
         headerName: "Used",
         width: 100,
+      },
+
+      {
+        field: "actions",
+        type: "actions",
+        headerName: "Actions",
+        width: 100,
+        cellClassName: "actions",
+        renderCell: (params) => {
+          return (
+            <div
+              className="d-flex justify-content-between align-items-center"
+              style={{ cursor: "pointer" }}
+            >
+              <MatEdit index={params.row.id} handleClick={() => this.props.history.push({pathname:"discount/edit",state:{id:params.row}})} />
+            </div>
+          );
+        },
       },
     ];
